Move the new-tab announcement into General_Block
General_Block is the home for block logic that belongs to no single
block, and this is that. Setup goes back to being registration only.
The constants, the filter and its tests move with it, unchanged.
