+++ src/utility.scss
@@ -1,13 +1,6 @@
 // Import tooltip format styles for frontend display.
 @use "formats/tooltip/style";
-// Hide text meant only for screen readers. Core ships `.screen-reader-text`
-// in the block-library stylesheet and most themes carry it, but neither is
-// guaranteed, and without it these notes render as visible text. Notes carry
-// this class as well, so they hide wherever they land, inside a block or a
-// tooltip in a paragraph. Declarations mirror core's, with `overflow-wrap`
-// standing in for its legacy `word-wrap` alias. No `:focus` counterpart:
-// nothing hidden this way is focusable, and adding one would fight a theme's
-// own skip-link styling.
+// Mirrors core's `.screen-reader-text`, which not every theme ships.
 .gatherpress--screen-reader-text {
 	position: absolute;
 	width: 1px;
